fix pfmlib_parse_event() to handle comma correctly
This was causing issues with the AMD64 symbolic PMU
names which have a comma (soon to be removed).

As a convenience, the parsing code stops processing
the event string at the first comma. This makes it
easier to pass list of events separated by commas.

This patch moves the check for comma until after the
PMU name has been checked.

Signed-off-by: Stephane Eranian <eranian@gmail.com>