fix unit mask combinations for generic cache events
We need to add umakss groups in order to enforce
the constraint that only one umasks for each of the
two groups can be selected at any one time.
Groups are:
 - perf_hw_cache_op_id
 - perf_hw_cache_op_result_id

Only one umasks of each group can be selected. This is
due to the way the encoding is done.

Signed-off-by: Stephane Eranian <eranian@gmail.com>