create distinct PMU structure for each AMD64 processor
In order to allow each individual event to be named without
running on the corresponding PMU, we expose each variations
of the AMD64 PMU by adding a pfm_pmu_t for each processor
revision (they all have distinct set of events). The patch
creates:
        [52, amd64_k7, "AMD64 K7"]
        [53, amd64_k8_revb, "AMD64 K8 RevB"]
        [54, amd64_k8_revc, "AMD64 K8 RevC"]
        [55, amd64_k8_revd, "AMD64 K8 RevD"]
        [56, amd64_k8_reve, "AMD64 K8 RevE"]
        [57, amd64_k8_revf, "AMD64 K8 RevF"]
        [58, amd64_k8_revg, "AMD64 K8 RevG"]
        [59, amd64_fam10h_barcelona, "AMD64 Fam10h Barcelona"]
        [60, amd64_fam10h_shanghai, "AMD64 Fam10h Shanghai"]
        [61, amd64_fam10h_istanbul, "AMD64 Fam10h Istanbul"]

This allows specific PMU event to be named:
$ showevtinfo amd64_k8_revf::cpu_clk_unhalted

The old PFM_PMU_AMD64 PMU identifier is now obsolete and should
not be used.

Signed-off-by: Stephane Eranian <eranian@gmail.com>