fix AMD64 HYPERTRANSPORT unit masks combination constraints
The umasks for HYPERTRANSPORT events are a bit tricky to combine
because of the SUBLINK_MASK umask. The ALL umask encapsulates all
the other umasks, except SUBLINK_MASK. SUBLINK_MASK may be provided
with every other umasks, incl. ALL.

This commit implements the constraints for all 4 HYPERTRANSPORT
events by using the grouping capability and introducing the FL_OMIT
flag to AMD64 events. It specifies that the umask can be omitted.
If all umasks in a group can be omitted and there is no default
for a group, then the library will not flag an error.

Signed-off-by: Stephane Eranian <eranian@gmail.com>