do not use arch_version to check for ATTR_T modifier
We must use the event's modmsk bitfield to determine
if the ATTR_T (arch version 3 or higher) is supported and
therefore must be appended to fstr. This is because we now
support operating on events that are not supported by the
host HW.

This patch also fixes  pfm_intel_x86_display_reg() for the
same reason. We make it private to pfmlib_intel_x86.c as it
is not used anywhere else.

Signed-off-by: Stephane Eranian <eranian@gmail.com>