unifiy ARCH_X86_64 and ARCH_I386
Make only on ARCH_X86. This simplifies Makefiles and
code in pfmlib_common.c.

Some on the processors compiled in 64-bit mode do not
actually support 64-bit mode, but that's okay. Their
event table will not be activated on 64-bit processor.
But that gives us a chance to validate the tables on
a 64-bit host systems.

Signed-off-by: Stephane Eranian <eranian@gmail.com>