remove pfmlib_getcpuinfo_attr() to gain portability
We remove pfmlib_getcpuinfo_attr() which was an OS
dependency and replace it with an actual call to cpuid()
on Intel X86. This way, the library becomes more portable.
It compiles on Linux, Mac OS X, FreeBSD.

On ARM, there is no equivalent to cpuid() that can be called
from user space. So instead, we have kept pfmlib_getcpuinfo_attr()
there to peek at /proc/cpuinfo. The function is private to ARM. The
code also relies on a custom implementation of getline() called
getl(). That also removes a dependency on this GNU libc extension
which may not always be present in embedded environments.

Signed-off-by: Stephane Eranian <eranian@gmail.com>