do not decompose reg.sel_umask to get umasks strings
In the  generic intel x86 and nhm/wsm uncore encoding
routines, we were decoding the umask  field, once we had
it encoded to build the fully qualified event string.
For that we also had to have a grpmsk field in the event tables
to help disambiguate umask codes.

Instead of all this complexity, we build the fully qualified
string as we encode umasks, incld. the default umasks.

We lose the fact that umasks were always listed in the same
order regardless of how they were specified in the event string.
This is a minor problem (when trying to automate fqstr comparisons)
but this will be fixed in a subsequent patch.

Signed-off-by: Stephane Eranian <eranian@gmail.com>