sort attributes to guarantee fstr remains constant
If the umasks are given is a different order, the fstr
should be the same. That was not the case since commit
a66e19a. Fix this by sort the attributes by id and then
generating the fstr. Do for all X86 arch. Need to do the
others which have umasks as well.

Signed-off-by: Stephane Eranian <eranian@gmail.com>