fix AMD64 invalid encoding of events with umask revision constraints
The encoding was not picking up the right umask, e.g., for
AMD K8 PROBE:ALL for instance. There is a ALL encoding up till
version C, then another one from D-G.

That was coming from a bogus amd64_umask_valid() and also because
both pfm_amd64_get_event_info() and pfm_amd64_get_event_attr_info()
were relying on attribute indexes being contiguous. They are not
because some umask may be skipped.

The common code was also at fault because it was not using the
ainfo.idx to encode the index of the attribute.

The pfmlib_sanitize_event() routine must look at the attribute
type to find duplicates.  Two attributes are duplicates if they
have the same attr index AND the same type. On AMD64, due to
the gaps in umasks, a umask and a modifier may have the same
index.

Signed-off-by: Stephane Eranian <eranian@gmail.com>