add a simple validation tests
There is a new subdir (tests) which contains a basic validation
test suite for the event tables and encodings. It supports a
generic test + x86 specific event test at the moment. It will be
extended to other architectures.

To run the test:
	validate -a

The minimal validation done is showevtinfo has been removed.

Signed-off-by: Stephane Eranian <eranian@gmail.com>