Finally fix the -WExtra problem.
The issue was -WExtra was being passed to libpfm, but only in the case
  where the user had a CFLAGS env variable.

It turns out this is due to the following from section 5.7.2 of the
gmake manual:
   Except by explicit request, make exports a variable only if it is
   either defined in the environment initially or set on the command line,

And the fix is also described:
   If you want to prevent a variable from being exported, use the
   unexport directive,

So I've added an "unexport CFLAGS" directive, which seems to be the
right thing as our Makefile explicitly passes CFLAGS to the
sub-Makefiles that need it.  This seems to fix the build.