The multiplex code currently does not make a final adjustment at the time of MPX_read().  This is to avoid the case where counts could be decreasing if you have multiple reads returning estimated values before the next actual counter read.
While this code works to keep the results non-decreasing, it can cause significant
differences from expected results for final reads, especially if many counters are
being multiplexed.

This is seen in the sdsc-mpx test.  It was failing occasionally on some machines
by having error of over 20% (the cutoff for a test error) when multiplexing 11
events.

What this fix does is to special case the PAPI_stop() case when multiplexing is
enabled, having the PAPI_stop() do a final adjustment.  The intermediate
PAPI_read() case is not changed.

This fixes the sdsc-mpx case, while still passing the mendes-alt case (which
checks for non-decreasing values).

There is a #define that can be set in multiplex.c to restore the previous behavior.