fix parsing of event string with comma
An recent commit borken parsing of event string
with comma, such as: core::rat_stalls:u:c=1,cycles.
libpfm allows the comma, it simply ignore everything
after the first comma. The problem is that current
code was working for events specified without attributes.

This patch fixes the issue by hoisting the lookup for
the comma at the very beginning of the function. This is
now safe because the library guarantees that no PMU name
(pname) have commas in them.

The patch also includes a validatation test for this
situation.

Signed-off-by: Stephane Eranian <eranian@gmail.com>