drop ptrace() for enable_on_exec+pipes
That's the easiest solution and it avoids the time tracking
issues with time_enabled.

With ptrace(), time_enabled starts when the event is created
even though the task has not actually executed a single instruction.
To avoid this, an event would have to be created disabled, then an
explicit IOC_ENABLE would have to be used.

Here we use the other approach which is based on the enable_on_exec
feature and pipes to synvchronize between the child and the parent.

Signed-off-by: Stephane Eranian <eranian@gmail.com>