Add cpuid leaf4 cache detection support.
This has been available on intel processors since Late model P4s
and all Core2 and newer.  It returns cache info in a different way
than the older leaf2 method.

Currently we only use leaf4 data if the leaf2 results tell us to
(apparently Westmere does that).  Otherwise we use the old method.

It might be interesting to use more of the leaf4 info.  It can tell
us things such as how many processors share a socket, how many
processors share a cache, and info on the inclusivity of a cache.