Change sanity check logic to apply to non detected PMUs
Call pfmlib_sanity_checks() even on PMU models which have not
been detected. That implies that certin entries in pfmlib_pmus[]
may not have mapping in  pfmlib_pmu_map[] so take care of that as
well. Introduce a PFMLIB_PMU_FL_INIT to tag PMU models which were
initialized properly.

Signed-off-by: Stephane Eranian <eranian@gmail.com>