Fix dependency on -ldl
Now configure checks if dl* symbols are in the base system libraries (i.e., no -ldl needed). If so, avoid adding -ldl to shlib example.

If dl* symbols are not find in the base system libraries, then check for -ldl, and if it exists, pass it to ctests/Makefile through Makefile. If -ldl is not found, fail at configure time.