change meaning of INTEL_X86_PEBS for events
We change the logic slightly to make it more convenient to detect
PEBS events.

In the past, if only some umasks supported PEBS, then the event flags
would not set INTEL_X86_PEBS, only the umasks would.

With this patch, the event sets INTEL_X86_PEBS when it has at least
one umask which supports PEBS or it has not umasks and supports PEBS.
Each umask supporting PEBS must set INTEL_X86_PEBS explicitely.

This patch make the changes to implement this change in logic.

Signed-off-by: Stephane Eranian <eranian@gmail.com>