drop event equivalence for fixed counter events
That was causing too much confusion and difficulties
when trying to enumerate event attributes.

Depending on the kernel interface, there is no guarantee
that any of the fixed counter events end up on fixed counters.
Some of those events work the same way on generic counters,
except  unhalted_reference_cycles.

Signed-off-by: Stephane Eranian <eranian@gmail.com>