drop perf_event.h
Instead of maintaining our own version of perf_event.h, use
the one provided by the host distribution.

You need to have the linux kernel header package installed
And kernel needs to be 2.6.31 or later. This is the case for
most distributions these days. On Ubuntu, the package is called
linux-headers-generic.

Signed-off-by: Stephane Eranian <eranian@gmail.com>