change the logic struct the size field in structs passed to the library
The size field is used to provide extensibility of the structs while maintaining
backward compatibility. It is more flexible than using reserved fields. You may
run out of those.

Each struct has a size field. It needs to be initialized to sizeof(struct xxx)
where xxx is the struct being passed. If size = 0, then the library considers
the user is passing the reference struct, i.e., the first ABI version of the
struct.

This mechanism detects any mismatch of a struct passed by user. Note that it
is okay to pass a struct that is bigger than what the library understands
but the extra bits need all be zero. In other words, struct must be memset()
then struct->size must be set.

The struct impacted by the change are:
 - pfm_pmu_info_t
 - pfm_event_info_t
 - pfm_event_attr_info_t
 - pfm_pmu_encode_arg_t
 - pfm_perf_encode_arg_t

The reference sizes are provided in constants for both 32 and 64 ABIs.

Signed-off-by: Stephane Eranian <eranian@gmail.com>