Update the POWER7 event set to the latest event specification.
A number of new events have been added, and a number of existing
events have been updated with corrected descriptions.  Also, many
events have been renamed to be more consistent and make it easier
on the tools.

A previous version of the automatic code generation tool corrupted
the event descriptions when the description contained one or more
commas.  This led to incorrect descriptions for many events.  This
patch fixes that issue.

Since this file is automatically generated, the patch is not minimal,
and in addition to that, some of the events whose description has
been corrected, moved relative to where they used to be.

Signed-off-by: Corey Ashford <cjashfor@linux.vnet.ibm.com>