Set size field before calling pfm_get_perf_event_encoding()
Anton Blanchard discovered that when he ran a simple example of
evt2raw on a POWER7 machine, he got a seg fault in libpfm because
libpfm checked the size field of the perf_event_attr struct, and
if it's non-zero, it uses that size to memset the struct to all
zeros.  If the size is garbage (a large number), memset will write
past the end of the structure, corrupting memory, and possibly
causing a seg fault.

This patch corrects that by initializing the size field to the
size of the perf_event_addr_struct which is being passed in.  This
should be the safest route.

Reported-by: Anton Blanchard <anton@au1.ibm.com>
Signed-off-by: Corey Ashford <cjashfor@linux.vnet.ibm.com>