reduce size of the Intel X86 event tables
The goal of this change is to minimize the space needed for
the Intel X86 event tables.

With the current code, the Intel X86 events were defined as:

struct intel_x86_event {
	unsigned int code
	unsigned int numasks;
	intel_x86_umask_t umasks[MAX_UMASK];
};

Turns out that in general events have few umasks, except for
a handful of events, such as offcore_response on NHM/WSM/SNB.
To accomodate those few events, the umasks array had to sized
to the largest number of umasks. The consequence was that a
lot of space was wasted.

With this patch:

struct intel_x86_event {
	unsigned int code
	unsigned int numasks;
	intel_x86_umask_t *umasks;
};

The umask table is defined separately from the event itself.
Thus, only the space actually needed for an event is allocated.

All the Intel X86 event tables had to be rewritten to this new
format. This transformation was done programmatically. An
advantage of this approach is that now all the event tables are
formatted the same way.

The size reduction is significant: 50%.

The patch also modifies the way the ucode field is encoded.
The umask code is now encoded at its final position in the register,
bit 8-15.

Signed-off-by: Stephane Eranian <eranian@gmail.com>