Add support for handling multiattach properly.
This adds a pid argument to the _papi_hwi_lookup_or_create_thread()
call.  A pid of "0" falls back to the old behavior of using
the current tid/pid.

If attaching to an outside pid/tid, a new thread object is created
to handle this.  This seems like the right thing to do, though
there's enough complicated code in the threads code that I haven't
fully audited that this can't fail somehow in complicated cases
where lots of attaching/detaching is done in conjunction with
having a large multi-threaded program.