For libpfm4 we were setting available counters to the number of generic counters.
This was less than libpfm3, so update the code to set the number of
counters to be equal to generic+fixed.

In theory whether an event can be added is determined at add time,
so the extra check for number of counters is unnecessarily getting
in the way.  This should be fixed but might require a re-write
of some PAPI internals.