Add remaining changes needed for ARM compilation. This is enough for "papi_avail" and "papi_native_avail" to work.
Lots of #warning statements scattered around.  ARM is a complicated
architecture and things like memory barriers and mutexes are very
dependent on what version of the architecture they are running on.
It will take a while to figure out the proper way to handle this in
PAPI.

Also, on Cortex-A8 and Cortex-A9 there is no way to separate kernel events
from the user ones.  So all measurements contain both.  This will probably
confuse our ctests.