Fix ARM event encoding
The below patch is needed for PAPI to work.  I'm not sure why that shift
was in there but it looks like it is unncessary.  With the patch PAPI
returns sane results for a lot of our tests (things are difficult because
you only get counts if exclude_user, exclude_kernel, and exclude_hv are
all set to 0).

Signed-off-by: Vince Weaver <vweaver1@eecs.utk.edu>