correctly set unsupported plm in pfm_get_perf_event_encoding()
When a certain priv level is not supported by the underlying hw,
it's exclude_* field must be set to zero (and not 1).

Introduce a supported_plm field in each pfmlib_pmu_support_t struct
to export a bitmask of supported priv level masks.

For the generic PMU exposed by perf_event, we use a best effort
approach by trying to figure out the actual priv levels based on
the first core CPU PMU found.

This patch solves the problem that on ARM (and possibly other
architectures), libpfm was incorrectly setting exclude_hv = 1
causing the kernel to fail on event creation.

Signed-off-by: Stephane Eranian <eranian@gmail.com>