ctests, ftests, utils, and the component tests were all using some files in ctests.
These weren't being built when --with-no-cpu-counters was enabled,
so the PAPI build was breaking when that was enabled as well as a component.

Move the shared files to their own directory, testlib
Then update all the users to look in the right place.

After this commit you might need to do a "cvs -d update" to make sure
you get the new subdirectory.