change encoding for Intel unhalted_reference_cycles event
This event only exists on Fixed Counter 2, as such it does not
have an event encoding.

This patch changes the encoding to match what is expected by
perf_event (future LKML patch), so it can be scheduled on the
proper fixed counter.

Event code 0x00 is not used by any other event on any CPUs.

Signed-off-by: Stephane Eranian <eranian@gmail.com>