Update the AMD Family 14h (Bobcat) pre-defined events.
It turns out they are different enough from 10h that they need
their own category.

In going through the Fam14h BKDG it turns out that Bobcat has a really
nice set of events available, especially for Floating-Point/SSE
but also memory bandwidth.

With this change, all of the ctests pass on a Bobcat machine.