The coverity checker was reporting we forgot to fclose() /proc/cpuinfo in papi.c
The bigger question, is why were we unconditionally trying to open
/proc/cpuinfo in generic code in papi.c anyway?

Turns out it was to set the event masks properly for itanium and p4.

The platform code sets CPU vendor and family for us though, so if
we just make the event mask code use those values then we don't
have to open cpuinfo.  This also means that non-Linux users
with the misfortune of running on a P4 might actually work too.