fix handling of default umasks with Intel umodel tag
There was a problem in pfm_intel_x86_add_defaults() where it was
looking at the wrong list of umasks to add defaults. It must use
the list of possible attributes and not the full list of attributes.
Some umasks may be specific to certain PMU models, e.g., NHM vs. NHM-EX.

Also fix the X86 validation to authorize multiple default umasks per
group as long as they are for different PMU models.

Signed-off-by: Stephane Eranian <eranian@gmail.com>