fix encoding of offcore_response:any_response
The encoding of any_response was wrong when shown via
event listing. The actual encoding via the encoding routine
was however correct when any_response was specified on the
cmdline. If any_response was implicitely added by libpfm4,
then the encoding was incorrect.

Fix validate_x86.c to have the correct encoding

Signed-off-by: Stephane Eranian <eranian@gmail.com>