Moved cpu and inherit bits to end of structure for compat across all 4.x lines. Found by Will Cohen.
As it turns out, I ended up reviewing the CPU_ATTACH changes; I had not done
so before. This functionality actually belongs in PAPI_set_granularity. A CPU
is a natural unit of granularity of counting, and that value was speced in
papi.h a long time ago. Right thing to do here is leave the current attach stuff
but make it work as part of set_granularity.

Consider that a TODO for 4.3.