There has been some ongoing speculation about what would happen if you enabled Multiplexing and Overflow at the same time.
It turns out (at least on perf_events) that if you have kernel multiplexing,
the results are what you expect.  You get overflows, but less than in the
non-multiplexing case because the overflow counter isn't being run all the time.

The results for software multiplexing involved a segfault.  This is because
in the software multiplexing case the primary EventSet is a fiction; a set
of shadow EventSets are created behind the scene, and these are the ones used.
Therefore when you enable overflow, the overflow event is attempted to be
enabled on the fictious main EventSet.  There are no native events mapped
for it, so overflow tries to access native event array index "-1" which
causes bad things to happen.

This change avoids the issue by catching the "-1" case and failing accordingly.
We should probably decide if we want to catch the oflo/mpx
combination earlier and outright ban it.

I also went through a lot of the code involved adding comments, as it
was really hard following what was going on.  This involved the
infamously dense "_papi_hwi_remap_event_position()" function too.