The multiplex1_pthreads test was reporting a memory leak.
This is because the test was calling PAPI_unregister_thread() without
destroying its EventSets.

This added change adds code that at unregister_thread time will destroy
any events belonging to that thread.

This works on all the current ctests but I should check some of the
various corner cases not currently tested.