The Solaris build was broken.
This was because we were using PAPI_MAX_COUNTER_TERMS to walk the
predefined events "pos" array, when it is declared as size MAX_COUNTER_TERMS.

On x86 and most other architectures PAPI_MAX_COUNTER_TERMS >= MAX_COUNTER_TERMS
so it didn't matter, but on Solaris the reverse was true and it segfaulted.

I tried my best to audit the code to avoid a regression this close to release.