fix event enumeration on AMD64 processors
There was a bug in the initialization of AMD64 event tables
which caused problem on AMD64 K8, Fam10h tables.

The number of events was calculated at runtime to match the
actual host processors. With Fam10h, the table contains events
which may be specific to a particular processor revision. Thus,
the table may contain a superset of the events supported by an
actual processor. However pfm_amd64_pmu_init() was overriding
the pmu->pme_count field. Yet that was used to iterate over ALL
the events in the table. Events for specific processor revision
may be scattered throughout the table. Thus, ALL event indexes
must be scanned. On Fam10h, the full table contains 129 events,
but Rev B (Barcelona) only supports 119. But the 119th Barcelona
event may be at index 125 in the table.

The patch fixes the problem by not recomputing pmu->pme_count
at runtime.

The patch introduces a new internal API get_num_events() which
is optional and used by pfm_get_pmu_info() to return to the end-user
the actual number of events available for a specific processor.

Users MUST use iterators to scan all the events available. Events
indexes are NEVER guaranteed contiguous.

Signed-off-by: Stephane Eranian <eranian@gmail.com>