Modify Intel Sandybridge PAPI_FP_OPS and PAPI_FP_INS events to not count x87 fp instructions.
The problem is that the current predefines were made by adding 5 events.
With the NMI watchdog stealing an event and/or hyperthreading reducing
the numbr of available counters by half, we just couldn't fit.

This now raises the potential for people using x87-compiled floating point
on Sandybridge and getting 0 FP_OPS.  This is only likely if running
a 32-bit kernel and *not* compiling your code with -msse.

A long-term solution might be trying to find a better set of FP predefines
for sandybridge.