This changes fixes two race conditions that are probably the cause of the pthrtough double-free error.
When freeing a thread, we remove and free all eventsets belonging to that thread.
This could race with the thread itself removing the evenset, causing some
ESI fields to be freed twice.

The problem was found by using the Valgrind 3.8 Helgrind tool

  valgrind --tool=helgrind --free-is-write=yes ctests/pthrtough

In order for Helgrind to work, I had to temporarily modify PAPI to use
POSIX pthread mutexes for locking.  Is there any reason we don't use
these all the time?