Add support for reporting if we are running in a virtualized environment to the PAPI_hw_info_t structure.
This currently only works on x86.  it works by looking at bit 31 of
ecx after a cpuid (the "in a VM" bit) and then using leaf 0x40000000
to get the name of the VM software (this works for VMware and Xen at least)

x86_cache_info.c was renamed to x86_cpuid_info.c to better reflect what
goes on in that file (it does various things based on the cpuid instruction).

the testlib header was updated to report virtualization status in the papi
header (printed for things like papi_native_avail).