Have separate concept of "compiled-in" versus "active" components
With this change, the _papi_hwd[] component info array only contains
a null-terminated list of _active_ components.  The
_papi_compiled_components[] array has the original full list.

At init_substrate[] time a pointer to a component is only put in the
_papi_hwd[] list if it is successfully initialized.

In addition the PAPI_num_compiled_components() and
PAPI_get_compiled_component_info() calls have been added, but
this is probably a confusing interface so they might only
be temporary additions.