Fix issue when using more than 32 multiplexed events on perf_event
On perf_event we were setting num_mpx_cntrs to 64.

This broke, as the MPX_EventSet struct only allocates room
for PAPI_MPX_DEF_DEG events, which is 32.

This patch makes perf_event use a value of 32 for num_mpx_cntrs,
especially as 64 was arbitrarily chosen at some point (the actual
value perf_event can support is static, but I'm pretty sure
it is higher than 64).