Clean up ->get_virt_usec()
It no longer needs to be passed a context, so remove that from all callers.

Also, ->get_virt_cycles() was just a get_virt_usec()*MHz on most
platforms.  While this is a bit dubious (especially as MHz can't
be relied on) make this a common routine that will be added
at innoculate time if ->get_virt_cycles() is set to NULL.