Fix build when not using POSIX timers
The PAPI build system was being overly clever with how it defined
what kind of wall clock timers were to be used, so of course I broke
things when breaking the timer code out to make it a bit more
understandable.

This patch breaks out the timer define into two pieces; one
saying it's a POSIX timer and one saying whether to use HR timers
or not.