Add new PAPI_enum_cmp_event() function
This will be needed when we remove the 16-component limit.

Currently in PAPI_enum_event() the component number is gathered from
bits 29-26 of the eventcode.  This won't work anymore once we remove
those bits.

Also update the various components to not use PAPI_COMPONENT_MASK()
as this too will go away in the transition.