Fix PAPI event enumeration inside of VMware
VMware disables the availability of architectural events
when virtualized PMU is not available.

libpfm4 was checkign this when enumerating events, and we
would end up in the situation where ix86arch was marked
active but 0 events were available.

We didn't check for this error condition and thus end up
thoroughly confused.