clarify fstr usage model for pfm_get_os_event_encoding()
There is a bug in the definition of libpfm4 ABI version 0
for the fstr fields part of the pfm_pmu_encode_arg_t and
pfm_perf_encode_arg_t structs. The field is defined as
char **, when it should be char *.

We cannot fix this without modifying the ABI and we want
to avoid that.

This patch demonstrate how, despite the bug, the field
can still be used to retrieve the fully qualified event
string if needed.

   char *fstr = NULL
   pfm_pmu_encode_arg_t arg;
   arg.fstr = &fstr;
   ret = pfm_get_os_event_encoding("event",
                                   PFM_PLM0|PFM_PLM3,
                                   PFM_OS_NONE,
                                   &e);
   if (ret == PFM_SUCCESS) {
      printf("fstr=%s\n", fstr);
      free(fstr);
   }

This is cumbersome but works.

The patch updates check_events.c to use the
pfm_get_os_event_encoding() function.

The patch also updates the man page for pfm_get_os_event_encoding()
to provide the above code snipet as example.

The perf_util.c utility module is updated to capture fstr. It is
not used just yet by the perf_event examples.

Signed-off-by: Stephane Eranian <eranian@gmail.com>