Rewrite and merge of the nVidia Management library component.
This component attempts to expose all supported 'performance counters'
on each card cuda knows about at runtime. Much like the cuda component
reads happen on the card you're executing on at PAPI_read time.

The test included is a copy of the cuda helloworld test, but it
attempts to start/stop the event on each gpgpu. If you select an
event that is not supported on the card you're running on we should
fail gracefully but this has not been tested.