Update hw_info_t CPU frequency reporting.
Previously PAPI reported "float mhz" and "int clock_mhz".
In theory the first was the current CPU speed, and the latter
was the resolution of the cycle counter.  In practice they
were both set to the same value (on Linux, read from /proc/cpuinfo)
and not very useful when DVFS was enabled, as the value reported
was usually lower than the actual frequency running once CPU started
being used.

This change adds two new values "cpu_max_mhz" and "cpu_min_mhz"
which are read from
  /sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_max_freq and
  /sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_min_freq
if they are available, and falls back to /proc/cpuinfo otherwise.

All of the tests were updated to use cpu_max_mhz.

The old mhz and clock_mhz values are left for compatability reasons
(and set to cpu_max_mhz) but are currently unused otherwise.