Remove the 16-component limit
This turned out to be easier than I thought it would be.

Now determining which component an event is in is a two step process.

Before, the code shifted and masked to find the component from bits 26-30.

Now, _papi_hwi_component_index() is used.  There's a new native
event table which maps all native events (which are allocated
incrementally at first use starting with 0x4000000) to two values,
a component number and an "internal" event number.