Re-enable PAPI_event_name_to_code() optimization
In PAPI_event_name_to_code() there was a commented-out optimization
where we would check if an event name begins with "PAPI_" before
searching the entire preset list for an event name.

The comment says we had to disable this due to "user events", but
a check shows that was introduced in e7bd768850ecf90 and that
the "user events" it means is not the current support, but the now-removed
PAPI_set_event_info() function where you could change the names
of presets on the fly (even to something not starting with PAPI_).

Since we don't support that anymore, we can re-enable the optimization.