init supported_plm for perf PMU dynamically
The perf PMU is the geenric perf_evnet PMU.
it corresponds to all the perf_event hardware
events such as cycles, instructions. This is
a pseudo PMU. But when it comes to priv levels,
the behavior across physical PMU changes. The
kernl layer does not respond in the same way to
unsupported priv levels. Libpfm does not let users
set priv levels that do not exist int he actual HW.

That means that the perf PMU needs to "fish" the
list of supported plm from the underlying core PMU.
This was done in validate_pattrs() but not during encoding
leading to generic perf events causing errors on ARM
for instance: task -e cycles ... would returen EPERM.

This patch fixes the problem, by extending the init()
callback of the perf PMU to overwrite its default
supported_plm bitmask with the one from the underlying
core PMU, assuming there is one. That solves the problem
on ARM, for instance.

Signed-off-by: Stephane Eranian <eranian@gmail.com>