enable encoding on Nehalem uncore events
This is based on the LKML patch series
posted here https://lkml.org/lkml/2012/6/12/31.
We are fairly condifent that is the way things
are eventually going to be published and in particular
how how discovers the attr->type for the uncore PMU.

Note that this patch does not completely enables encodings
as the pmu->perf_name needs to be set appropriately in
the pfmlib_pmu struct first. This is addressed in a distinct
patch.

Signed-off-by: Stephane Eranian <eranian@gmail.com>