disambiguate modifier and umask with the same name
There was a problem when a umask and modifier had
the same name. The parser could not tell them apart.

This patch fixes this. The disambiguation uses the =
sign. For boolean modifiers, simply force the value
with i=0 or i=1: L2_LINES_IN:I:I=1 or, L2_LINES_IN:I:I=0

Add the tests to validate this as well.

Signed-off-by: Stephane Eranian <eranian@gmail.com>