Fix libpfm4 ntv_event_to_info event_info_t event value
The recently added libpfm4 ntv_event_to_info function was not
properly oring PAPI_NATIVE_MASK to the event value in the
event_info_t struct.  That means if you tried to use that event
value to add an event it would fail.

The overflow_allcounters test broke because of this.