Standardize component names.
This commit changes the name of a component as reported in its info struct
to be the name of the directory the component resides in.  The rational is
that at configure time you 'name' the components by the directory they reside
in, so we preserve this.

IE:
./configure --with-components="rapl"
My principle of least suprise would dictate the call should be
	PAPI_get_component_index("rapl");
	not PAPI_get_component_index("linux-rapl");