Remove "include CPUCOMPONENT" from papi_internal.h
This was the last major dependency on CPU component in common
PAPI code.  It was mostly necessary for the ucontext
definitions when trying to get the instruction pointer
when doing sampling.

This change creates new OS-specific header files that handle
the ucontext related code, and has papi_internal.h include
that instead.