Remove papi_vector.h include from papi_internal.h
There were some semi-circular dependencies that came up
with the context split changes.  The easiest way to
fix things for perfctr was just move papi_vector.h
out to be included explicitly.

This touches a lot of files because a lot of files
include papi_internal.h

This should also fix the perfctr and perfmon2 builds
that were broken yesterday.