Remove PAPI_MPX_DEF_DEG
It was not well documented and being used in confused ways all over the code.

Now there is a different define
  PAPI_MAX_SW_MPX_EVENTS
used by the software multiplex code.

All other components have had the value replaced with just the
maximum number of counters.  If a component can handle its own
(non-software) multiplexing it is up to it to set .num_mpx_cntrs
to a value that's different from .num_cntrs