perf_event: move all event specific info to the control state
previously half was in the context state and half in the control state

perf_event has a strange architecture with each event being created
having its own fd, which is context wide.  In PAPI though we usually
only have one eventset (control state) active at once, so there's no
need to have the context be aware of this.