perf_events: enable new read_code
This makes the read code much simpler.

It finishes the multiplexing changes.

To avoid complication, we no longer enable PERF_FORMAT_ID as reading
that extra info is unnecessary with the current implementation.

This passes all the tests on a recent kernel, but on 2.6.32 there
are still a few issues.